requirements.txt
gradio>=5.0.0
transformers>=4.40.0
torch>=2.2.0
numpy
pandas
matplotlib
wordcloud
Pillow
